+++ src/main/java/com/jnape/palatable/lambda/adt/hlist/Tuple2.java
@@ -76,6 +76,16 @@ public class Tuple2<_1, _2> extends HCons<_1, SingletonHList<_2>> implements
     }
 
     /**
+     * Snoc an element onto the back of this HList.
+     *
+     * @param _3   the new last element
+     * @return the updated HList
+     */
+    public <_3> Tuple3<_1, _2, _3> snoc(_3 _3) {
+        return tuple(_1, _2, _3);
+    }
+
+    /**
      * {@inheritDoc}
      */
     @Override